% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{pantherCompatibleEnsemlDatasets}
\alias{pantherCompatibleEnsemlDatasets}
\title{pantherCompatibleEnsemlDatasets}
\usage{
pantherCompatibleEnsemlDatasets()
}
\value{
dataframe with ensembl datasets represented in PANTHER
}
\description{
Lists all ensembl dataset compatible with latest PANTHER database
}
\examples{
datasets = pantherCompatibleEnsemlDatasets
}
