% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{run_PIP1}
\alias{run_PIP1}
\title{run_PIP1}
\usage{
run_PIP1(
  proteome_dir = file.path(system.file("extdata", package = "pip1"), "PD proteomes"),
  known_PD_genes = read_excel(file.path(system.file("extdata", package = "pip1"),
    "Known PD genes.xlsx")),
  output_dir = choose_directory("Choose output directory"),
  species_dataset = choose_dataset(),
  primary_PANTHER_classification = "subfamily",
  minimum_num_proteomes = 2,
  PIPAC_or_PIPBD_filter = "secreted_or_gpi_or_tmhmm",
  annotate_with_NCBI_geneInfo = FALSE
)
}
\arguments{
\item{proteome_dir}{path of directory containing proteomes: example datasets is set as default}

\item{known_PD_genes}{path of .xlsx containing experimentally verified PD proteins. example datasets is set as default}

\item{output_dir}{path where PIP1 will provide output: dialog will appear by default}

\item{species_dataset}{ensemble dataset to use; dialog will appear by default}

\item{primary_PANTHER_classification}{string: either "subfamily" or "family"}

\item{minimum_num_proteomes}{integer: Minimum number of proteomes for inclusion in list A and B}

\item{PIPAC_or_PIPBD_filter}{string ("secreted", "secreted_gpi_or_tmhmm", "secreted_or_gpi_or_tmhmm" , "tm_or_gpi", "gpi", "tm", "targetP_not_C_or_M", "targetP_not_C_or_M_or_dash") : logic choice to group output to PIPA or PIPC (if true) or PIPB or PIPD (if false). See associated publication.}

\item{annotate_with_NCBI_geneInfo}{bool: if true, will download NCBI data to get gene description and synonyms. Slow}
}
\value{
Returns null
}
\description{
Main function to generate an in silco proteome for a given species
}
\examples{
run_PIP1()
}
